requirements.txt
# Core data processing
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.10.0

# Data acquisition
yfinance>=0.2.28

# Machine learning - sklearn
scikit-learn>=1.3.0

# Machine learning - boosting
xgboost>=1.7.0
lightgbm>=4.0.0

# Deep learning
torch>=2.0.0
pytorch-lightning>=2.0.0

# Time series
statsmodels>=0.14.0

# Hyperparameter optimization
optuna>=3.0.0

# Technical analysis
ta>=0.10.0

# Optional: Text/NLP
transformers>=4.30.0  # For FinBERT
sentence-transformers>=2.2.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# Utilities
pyyaml>=6.0
tqdm>=4.65.0
python-dotenv>=1.0.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Jupyter
jupyter>=1.0.0
ipykernel>=6.25.0
